+++ cogs/fun.py
@@ -8,6 +8,8 @@ from library.ConsoleLib import Time
 from library.ConsoleLib import Essentials
 from library.ConsoleLib import Roundtrip
 
+from library.PoliticalSelect import States
+
 class Fun(commands.Cog):
     def __init__(self, client):
         self.client = client
@@ -19,7 +21,7 @@ class Fun(commands.Cog):
         answer = Essentials.CoinToss()
         await message.edit(content = f'{answer}')
 
-        print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE: COINTOSS - LOG: cointoss was utilized! [ Coin: {answer} ]')
+        print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE: COINTOSS - LOG: cointoss was utilized in #{ctx.channel}! [ Coin: {answer} ]')
    
     # PYTHON METHOD
     @commands.command( help = "The language it was programmed in.", aliases = ['py'] )
@@ -67,7 +69,7 @@ class Fun(commands.Cog):
         await message.delete()
         await ctx.reply(embed = embed)
         
-        print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE: EIGHTBALL - LOG: 8ball was utilized! [ Question: {question} ] [ Answer: {answer} ]')
+        print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE: EIGHTBALL - LOG: 8ball was utilized in #{ctx.channel}! [ Question: {question} ] [ Answer: {answer} ]')
     
     # POPPY METHOD
     @commands.command( help = 'You know what I am talking about.', aliases = ['hitler','fascist'] )
@@ -102,7 +104,25 @@ class Fun(commands.Cog):
             url = IMAGE
         )
         await ctx.reply(embed = embed)
-        
+    
+    @commands.command(help = 'Returns state flags. (information coming soon...)')
+    async def state(self, ctx, *, name):
+        embed = discord.Embed ()
+        embed.set_image (
+            url = States.states(str.lower(name))
+        )
+        await ctx.reply(embed = embed)
+        print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE: STATE - LOG: State was utilized in #{ctx.channel}! [State Flag: {name}]')
+
+    @commands.command(help = 'Returns territory flags. (information coming soon...)')
+    async def territory(self, ctx, *, name):
+        embed = discord.Embed ()
+        embed.set_image (
+            url = States.territories(str.lower(name))
+        )
+        await ctx.reply(embed = embed)
+        print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE: TERRITORY - LOG: Territory was utilized in #{ctx.channel}! [Territory Flag: {name}]')
+
     # .
     @commands.command(name = '.', help = '.')
     async def dot(self, ctx):